# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/dev/test-runs/execute": {
      "post": {
        "operationId": "DevController_executeTestRun",
        "summary": "Esegue la suite di regression test API",
        "description": "Bloccato in NODE_ENV=production.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tutti i test superati"
          },
          "500": {
            "description": "Almeno un test fallito"
          }
        },
        "tags": ["dev"]
      }
    },
    "/api/dev/reset": {
      "post": {
        "operationId": "DevController_reset",
        "summary": "Reset dati e storage (solo dev/test)",
        "description": "Richiede header x-reset-confirm: true. Bloccato in produzione.",
        "parameters": [
          {
            "name": "x-reset-confirm",
            "in": "header",
            "description": "Deve essere \"true\"",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset completato con fixture"
          },
          "400": {
            "description": "Header mancante"
          },
          "403": {
            "description": "Non disponibile in produzione"
          }
        },
        "tags": ["dev"]
      }
    },
    "/api/dev/seed-large-pdf": {
      "post": {
        "operationId": "DevController_seedLargePdf",
        "summary": "Seed template grande con campi parziali + generazione PDF",
        "description": "Crea un template contrattuale multi-sezione con 21 campi. Compila solo 9 campi (sede, CF cliente, CAP, durata, importi IVA/totale, modalita pagamento e firme rimangono come {{placeholder}} nel PDF). Il file PDF generato viene mantenuto in storage. Bloccato in produzione.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "PDF accodato, IDs restituiti"
          },
          "403": {
            "description": "Non disponibile in produzione"
          }
        },
        "tags": ["dev"]
      }
    }
  },
  "info": {
    "title": "MAC Documents API",
    "description": "API per la generazione documentale basata su template Markdown.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}